Moved Platform and related classes to the TCDFx.Runtime namespace
